clase6-paginacion/requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy
flask_migrate